Remove unecessary headers, add android binaries, change tokenizers to new decode API

+++ packages/react-native-executorch/common/rnexecutorch/TokenizerModule.h
@@ -11,13 +11,13 @@ class TokenizerModule {
 public:
   explicit TokenizerModule(std::string source,
                            std::shared_ptr<react::CallInvoker> callInvoker);
-  [[nodiscard("Registered non-void function")]] std::vector<int32_t>
+  [[nodiscard("Registered non-void function")]] std::vector<uint64_t>
   encode(std::string s) const;
   [[nodiscard("Registered non-void function")]] std::string
-  decode(std::vector<int32_t> vec, bool skipSpecialTokens) const;
+  decode(std::vector<uint64_t> vec, bool skipSpecialTokens) const;
   [[nodiscard("Registered non-void function")]] std::string
-  idToToken(int32_t tokenId) const;
-  [[nodiscard("Registered non-void function")]] int32_t
+  idToToken(uint64_t tokenId) const;
+  [[nodiscard("Registered non-void function")]] uint64_t
   tokenToId(std::string token) const;
   [[nodiscard("Registered non-void function")]] std::size_t
   getVocabSize() const;
